backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlmodel>=0.0.16
python-dotenv>=1.0.0
garminconnect==0.2.8
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
ollama>=0.1.6,<0.2.0
httpx>=0.25.0
pydantic>=2.0.0
